back/auth_service/requirements.txt
altgraph
asgiref==3.8.1
astroid==3.3.8
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
dill==0.3.9
Django==4.2.19
django-cors-headers==4.7.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
whitenoise==6.7.0
Faker==24.11.0
exceptiongroup==1.2.2
flake8==7.1.2
future
gunicorn>=20.1.0
idna==3.10
iniconfig==2.0.0
isort==6.0.0
macholib
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pylint==3.3.4
pylint-django==2.5.5
Faker==30.3.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
pytest==8.3.4
pytest-cov==6.0.0
requests==2.32.3
six
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
